+++ src/components/BoardView.jsx
@@ -3,7 +3,7 @@ import CourtCard from "./CourtCard.jsx";
 import QueueList from "./QueueList.jsx";
 import SectionLabel from "./SectionLabel.jsx";
 
-export default function BoardView({ state, waitingPlayers }) {
+export default function BoardView({ state }) {
   return (
     <div>
       <div style={styles.courtGrid}>
@@ -12,7 +12,7 @@ export default function BoardView({ state, waitingPlayers }) {
         ))}
       </div>
       <SectionLabel>Waiting queue</SectionLabel>
-      <QueueList waitingPlayers={waitingPlayers} />
+      <QueueList queueIds={state.queueIds} players={state.players} nextMatchups={state.nextMatchups || []} />
     </div>
   );
 }